src/test_langchain/pyproject.toml
[project]
name = "test-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.6",
    "langchain-anthropic>=1.3.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "langchain-google-community[calendar]>=3.0.5",
]
